Cleanups.
Former-commit-id: d00e52b3a73e86929794d8778bbe0aaed363947c

+++ client/src/include/event.h
@@ -46,7 +46,7 @@ enum
 typedef struct key_struct
 {
 	/** If 1, the key is pressed. */
-	int pressed;
+	uint8 pressed;
 
 	/** Last repeat time. */
 	uint32 time;
